requirements.txt
# Core ML & data
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
joblib>=1.3

# Visualisation
matplotlib>=3.7

# Web app
streamlit>=1.33
fastapi>=0.110
uvicorn[standard]>=0.29

# NLP — claim extraction (optional but recommended)
# pip install spacy && python -m spacy download en_core_web_sm
spacy>=3.7

# Semantic similarity (optional, falls back to TF-IDF without it)
sentence-transformers>=2.6

# News retrieval
requests>=2.31

# Web scraping (optional — used when scrape=True)
newspaper3k>=0.2.8
beautifulsoup4>=4.12
lxml>=5.0

# Data validation for FastAPI
pydantic>=2.0
